requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
llama-index>=0.11.0
llama-index-llms-gemini==0.5.0
llama-index-agent-openai==0.4.12
llama-index-embeddings-huggingface==0.5.5
llama-index-vector-stores-qdrant==0.6.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
qdrant-client==1.15.1